api/webapp/requirements.txt
gunicorn
Flask
Jinja2
Werkzeug
Flask-Cors==3.0.9
mysql-connector-python==8.0.5